+++ server/fd_handler.cpp
@@ -13,7 +13,7 @@ int count_nl (char *str,int len)
     return nl;
 }
 
-void servers::client_req_handler(int &index)
+void servers::client_req_handler(size_t &index)
 {
     if (!(fd_poll[index].revents & POLLIN))
         return ;
@@ -35,14 +35,14 @@ void servers::client_req_handler(int &index)
     }
 }
 
-void servers::client_res_handler(int &index)
+void servers::client_res_handler(size_t &index)
 {
     if (!((fd_poll[index].events & POLLOUT) && (fd_poll[index].revents & POLLOUT)) && !((fd_poll[index].events & POLLIN) && (fd_poll[index].revents & POLLIN)))
         return ;
     data[index].response->generate_response(&(fd_poll[index]),&(data[index].request),&(data[index].time));
 }
 
-void servers::listener_handler(int &index)
+void servers::listener_handler(size_t &index)
 {
     if (!(fd_poll[index].revents & POLLIN))
         return ;
