+++ server/http/GET_Handler/get_handler.cpp
@@ -25,7 +25,7 @@ void http_response::GET_check_state()
                 headers["LOCATION"] = (request->get_path() + "/");
                 throw 301;
             }
-            for (int i = 0;i < conf.index.size();i++)
+            for (size_t i = 0;i < conf.index.size();i++)
             {
                 std::ifstream f(conf.root + "/" + conf.index[i]);
                 if (f.good())
